media: bcm2835-unicam: Initialise ret value
A loop has been added which checks the return condition of a statement
executed inside the loop. This leaves the code path making a comparison
against an uninitialised return variable, which could be anything on the
stack for the first run of the loop.

Initialise the ret value as a simple fix.

Removes the following warning:

drivers/media/platform/bcm2835/bcm2835-unicam.c: In function ‘get_first_supported_format’:
drivers/media/platform/bcm2835/bcm2835-unicam.c:788:18: warning: ‘ret’ is used uninitialized in this function [-Wuninitialized]
  for (j = 0; ret != -EINVAL && ret != -ENOIOCTLCMD; ++j) {

Fixes: 06c0df8b484b ("media: bcm2835-unicam: Rework to not cache the list of active fmts")
Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>