dt-bindings: dwc3: add quirk to disable USBHstInAutoRetry
This feature automatically requests retries of a burst that has CRC
errors in some of its packets instead of terminating the burst and
trying again later.

In a system with high latency, the host will terminate a bulk IN burst
that overflows the RX fifo with a transaction error instead of
automatically NAKing and retrying, which is counterproductive as it
will halt the endpoint completely.

Add a quirk to disable this behaviour.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>