mfd: rp1: Ignore INTSTAT when handling interrupts
The logic for turning level-triggered RP1 interrupts into MSI-Xs
(which are edge-triggered) can cause bits in the PCIe INTSTAT to read
back as zero momentarily, which has been seen to mislead the RP1 driver
and cause interrupts to cease. Avoid the problem (and improve
performance) by removing the INTSTAT reads.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>