usb: xhci: drop and add the endpoint context in xhci_fixup_endpoint()
Setting both the Drop and Add bits on the input context prevents
corruption of split transactions seen with a Designware USB3 host
controller.

This is a downstream feature that allows usbhid to restrict polling
intervals on devices connected to xhci controllers, and was only tested
on a VL805 which didn't complain about the fact the endpoint got added
twice.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>