drivers: media: pisp_fe: Remove null config
Remove the use of null config as it was only ever needed for early
driver development and bringup. Because of this change, ensure there
is always a config buffer when submitting a job.

Remove the blocking of AXI config writes from the config buffer, since
this is not done through the null config any more.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>