media: platform: rp1: copy and validate ISP-BE configuration
PISP-FE driver formerly overwrote some parts of the configuration
structure, but failed to check others. Now we take a copy and use
but validate the user-supplied parameters. Validation is needed
to ensure it won't clobber memory or lock up.

Also changed the logic that decides which sub-block configs are
written, to ensure output and crop blocks are always configured
when enabled, and avoid duplication.

Signed-off-by: Nick Hollinghurst <nick.hollinghurst@raspberrypi.com>