media: rkisp1: Prevent race when setting digital zoom on imx8mp
When modifying the sink crop and scaler for digital zoom on the i.MX8MP,
this will change parameters on both the ISP (for crop) and the resizer
(for scaling), and changing these two are not atomic. This could thus
race, so we catch this by detecting buffer wraparound, and mark the
buffer as error.

Signed-off-by: Paul Elder <paul.elder@ideasonboard.com>