media: platform: rp1: Convert dma_addr_t to u64 before using
dma_addr_t is an opaque type, the size of which can vary based on the
platform. If the dma_addr_t is 32 bits, the current code causes
compile warnings.

When writing the dma address to registers, convert the value first to
u64 so that we operate on a known type.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>