media: rkisp1: Fix 34-bit DMA addressing
The driver deals with the DMA addresses using u32. This has worked in
earlier versions of the IP, but IMX8MP ISP, the addressable space is 34
bits. If the system allocated CMA memory happens to be located higher
than 32-bits, the driver will fail to allocate CMA memory, afaics falls
back to kmalloc(), which will fail for larger allocations (over
MAX_ORDER, which often mean 4MB).

Fix this by setting the dma mask to 34 bits when allowed, and fixing the
code to use dma_addr_t when storing dma addresses, instead of u32.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>