drivers: media: pisp_be: Fix for job queue removal in stop_streaming()
The existing code unconditionally removes jobs from the job_queue list
when all the nodes in a node group have stopped streaming. This will
also remove jobs for any other node groups as the job_queue is a common
list. Fix this by only conditionally deleting jobs that belong to the
current node group.

Additionally, delete jobs as soon as the first node in the node group
stops streaming. Running a job with an incomplete set of active nodes
is invalid.

Fixes: 880153e0d0a9 ("media: pisp_be: Split jobs creation and scheduling")
Signed-off-by: Naushir Patuck <naush@raspberrypi.com>