media: i2c: imx283: Move binning to scan modes
The binning factors are determined by the chosen scan mode.

Move the definition of the binning ratio to the scan mode strutures
and remove from the v4l2 output mode definitions. The horizontal
binning ratio is not used and therefore is dropped.

This also fixes the 10-bit mode handling which previously had an
undefined vbin_ratio for MODE1.

V4L2 does not currently expose an API to support the differences between
binning and skipping, so while the mode capabilities are kept for the
skipping modes - there is no definition to use them yet.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>