media: i2c: imx283: Update exposure range on blanking changes
The exposure ranges are updated currently whenever the VBLANK control
is updated.

However, the total exposure limits are a factor of both the HBLANK and
the VBLANK durations.

Determine the hmax value any time the control is set, including
if the device is not yet streaming and update the limits accordingly.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>